refactor: tighten the decomposition's module boundaries and names
- Compact `binding`, `colon_targets`, and `legacy_union_syntax` to flat files under a three-file minimum for directory modules.
- Rename the `test_support` module to `testing`, matching the one-word names of its siblings.
