+++ src/primitives/imports.rs
@@ -66,7 +66,7 @@ mod tests {
     use rstest::rstest;
 
     use super::*;
-    use crate::test_support::parse;
+    use crate::testing::parse;
 
     #[rstest]
     #[case("import os\n", &[], Some(ImportGroup::Bare))]
