+++ src/rules/docstring_expand.rs
@@ -53,7 +53,7 @@ impl Rule for DocstringExpand {
 mod tests {
     use super::*;
     use crate::pipeline::Pipeline;
-    use crate::test_support::parse;
+    use crate::testing::parse;
 
     fn run(src: &str) -> String {
         let source = parse(src);
