+++ src/rules/docstring_wrap.rs
@@ -252,7 +252,7 @@ fn rewrite_body(
 mod tests {
     use super::*;
     use crate::pipeline::Pipeline;
-    use crate::test_support::parse;
+    use crate::testing::parse;
 
     fn run(src: &str) -> String {
         let source = parse(src);
